Also handle crypt device type in device stack
In what is probably the most common case, there is only the one crypt
device (the LUKS device itself) in the device stack, and everything
below it is non-crypt. However, there can be some configurations where
the LUKS device may have other crypt devices below it in the device
stack. Just one such example is when the LUKS device is a MD-RAID
array with a journal device that is also on a LUKS-encrypted
device. In such a case, the MD-RAID device which is a LUKS device, is
topologically dependent upon the journal device, another LUKS device.