Add support for BIOS
Most of the procedure is the same. But unlike UEFI, where grub-install
can determine the disk containing the EFI System Partition (ESP) based
on what's mounted at /boot/efi, on BIOS the user needs to tell us
which disk is the boot disk (which disk the BIOS will load the master
boot record from). Then, instead of copying the generated core image
to the ESP, on BIOS the core image is copied to /boot/grub/i386-pc and
then has to be installed from there to the boot disk (either to a BIOS
Boot partition on GPT, or embedded in the "gap" on MBR). The
grub-setup-bios utility exists for doing that installation.